HealthHub/requirements.txt
aiosmtplib==1.1.6
anyio==3.6.1
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
# Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
itsdangerous==2.1.2
messages==0.8.0
mysqlclient==2.1.0
# numpy==1.23.1
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
validus==0.3.0
